Adding Teacher component src
